[aml] git: use repository directly rather than archiving
Rather than cloning a repo, creating an archive of a revision, and extracting
that archive for each build, we can be much more efficient by working directly
from the git repository.

This uses a trick to "touch" a phony archive to mark the revision as
downloaded, then later ignoring it and just doing a checkout instead.