[pivos] fixed gl/egl libs
Because the libs have no SONAME set, the linker gets confused and sets the .so
as the required file. These may or may not be cleaned by various install scripts.

Instead, move the libs to .so and remove the symlinks. This should be fixed by
AMLogic.

Also, remove the corrpted glesv1 file.