[aml] HACK. Set HOME for init scripts.
There are probably 10 better ways, but this works for now.

Note that ~/.xbmc and others will move, so don't be alarmed by a fresh
XBMC profile.

This fixes things that depend on HOME being set properly. The particular issue
in this case was scrapers not loading due to the xmls in system addons not
being found.