[aml] do not wipe userdata for Linux->Linux updates by default
We test for the existence of a Linux file that would not be present on android
(/etc/issue). If that file exists, assume we're updating and do not wipe
userdata. This could later be expanded to only wipe if the contents of the file
contained a certain string.

The only way to achieve this in a reasonable way was to move to an ics-based
updater binary. I did not encounter any issues with the switch. The binary
comes from the ICS update image.

TODO: Build that binary from source.

NOTE: This was tested by faking an Android layout and verifying that userdata
is wiped under those conditions. A real Android->Linux wipe needs verification.