Remove patch that enabled write caching of usb drives and mSD Card
This is potentially the lesser of two evils.  There is no method in XBMC to enable the save removeal of USB drives.

With write cache enabled it can be up to 30 seconds from actual change, before the data is flushed to disk.
This can cause disk corruption for external USB drives on suspend of an M3, corruption of mSD card on a hard reboot, and can cause file systems to ultimately end up mounting read only.

Running with "sync" enabled on usb storage can in theory shorten life of storage device but is better than constant problem of disk corruption.  So removing the patch to change from "dirsync" to "sync" will cause usb devices and mSD Card to write straight to disk.