Patch to usbmount to allow for mountpoint specification
Mount by UUID or LABEL is currently not implemented in usbmount.
And both methods are often difficult for novice users to use.

This patch adds the idea of a ".mountpoint" file.
If file exists on removable media being mounted, and file contains name/label.
The media will be mounted in /media with the name in the file label.

For example if a hard drive is plugged in and is currently mounted as /media/usb0.
And if user wants to have drive always mount as /media/MyFiles, they can execute:
echo MyFiles > /media/usb0/.mountpoint
Then reboot and from then on the drive will moount as /media/MyFiles